requirements.txt
pywebview
requests